third_party/blink/web_tests/external/wpt/web-nfc/nfc_hw_disabled-manual.https-expected.txt

This is a testharness.js-based test.
FAIL NFCReader.start should fail if NFC HW is disabled. promise_test: Unhandled rejection with value: object "ReferenceError: NFCReader is not defined"
PASS NFCWriter.push should fail when NFC HW is disabled.
Harness: the test ran to completion.
